+++ test/test_bioinfo.rb
@@ -14,12 +14,9 @@ class Bioinfo_Test < Test::Unit::TestCase
   context "Bioinfo module" do
     should "have such hierarchy" do
       assert_nothing_raised do
-        # Module
         Bioinfo::Modules
         Bioinfo::Databases
         Bioinfo::Scripts
-        # Class
-        Bioinfo::Script
       end
       
       assert_raise NameError do